Update search.rst
There was no documentation about how to define the block (atleast not that I found). This resulted in an error when you tried searching. 

'An exception has been thrown during the rendering of a template ("The block type sonata.admin.block.search_result does not exist") in SonataAdminBundle:Core:search.html.twig at line 48'

There is a summary of the issue here:
http://stackoverflow.com/questions/19551715/sonata-admin-search-feature